Fix missing entity error
