+++ src/ui/components/ShipPanel.tsx
@@ -1,70 +1,10 @@
 import React from "react";
 import { Ship } from "../../archetypes/ship";
-import { MineOrder, Order, OrderGroup } from "../../components/orders";
 import { commodities } from "../../economy/commodity";
-import {
-  Collapsible,
-  CollapsibleContent,
-  CollapsibleSummary,
-} from "./Collapsible";
 import { Docks } from "./Docks";
-import { Sim } from "../../sim";
-import { asteroidField } from "../../archetypes/asteroidField";
 import AutoOrder from "./AutoOrder";
 import { Commander } from "./Commander";
-
-function getOrderDescription(ship: Ship, order: Order) {
-  switch (order.type) {
-    case "move":
-      return `Go to ${
-        ship.sim.getOrThrow(order.targetId).cp.name?.value ?? "location"
-      }`;
-    case "teleport":
-      return "Teleport to location";
-    case "trade":
-      return order.offer.type === "sell"
-        ? order.targetId === ship.cp.commander?.id
-          ? `Deliver ${order.offer.quantity}x ${order.offer.commodity} to commander`
-          : `Deliver ${order.offer.quantity}x ${order.offer.commodity} to ${
-              ship.sim.getOrThrow(order.targetId).cp.name?.value
-            }`
-        : `Get ${order.offer.quantity}x ${order.offer.commodity} from ${
-            ship.sim.getOrThrow(order.targetId).cp.name?.value
-          }`;
-    case "mine":
-      return `Mine ${
-        ship.sim
-          .getOrThrow(order.targetFieldId)
-          .requireComponents(["asteroidSpawn"]).cp.asteroidSpawn.type
-      }`;
-    case "dock":
-      if (order.targetId === ship.cp.commander?.id)
-        return "Dock at commanding facility";
-      return `Dock at ${ship.sim.getOrThrow(order.targetId).cp.name?.value}`;
-    default:
-      return "Hold position";
-  }
-}
-
-function getOrderGroupDescription(order: OrderGroup, sim: Sim) {
-  switch (order.type) {
-    case "move":
-      return "Move";
-    case "trade":
-      return "Trade";
-    case "mine":
-      return `Mine ${
-        asteroidField(
-          sim.getOrThrow(
-            (order.orders.find((o) => o.type === "mine") as MineOrder)!
-              .targetFieldId
-          )
-        ).cp.asteroidSpawn.type
-      }`;
-    default:
-      return "Hold position";
-  }
-}
+import Orders from "./Orders";
 
 const ShipPanel: React.FC<{ entity: Ship }> = ({ entity: ship }) => {
   const storedCommodities = Object.values(commodities).filter(
@@ -102,20 +42,7 @@ const ShipPanel: React.FC<{ entity: Ship }> = ({ entity: ship }) => {
           <hr />
         </>
       )}
-      {ship.cp.orders.value.length === 0
-        ? "No orders"
-        : ship.cp.orders.value.map((order, orderIndex) => (
-            <Collapsible key={`${order.type}-${orderIndex}`}>
-              <CollapsibleSummary>
-                {getOrderGroupDescription(order, ship.sim)}
-              </CollapsibleSummary>
-              <CollapsibleContent>
-                {order.orders.map((o, index) => (
-                  <div key={o.type + index}>{getOrderDescription(ship, o)}</div>
-                ))}
-              </CollapsibleContent>
-            </Collapsible>
-          ))}
+      <Orders ship={ship} />
       <hr />
       {!!ship.cp.docks && <Docks entity={ship.requireComponents(["docks"])} />}
     </div>
